Move isFirstClick to Visit model
