+++ app/Services/VisitService.php
@@ -38,26 +38,11 @@ class VisitService
         $visit->url_id = $url->id;
         $visit->user_type = $this->userService->userType();
         $visit->user_uid = $this->userService->signature();
-        $visit->is_first_click = $this->isFirstClick($url);
+        $visit->is_first_click = $visit->isFirstClick($url);
         $visit->referer = $this->getRefererHost($referer);
         $visit->browser = $deviceDetector->getClientAttr('name');
         $visit->os = $deviceDetector->getOsAttr('family');
         $visit->save();
-    }
-
-    /**
-     * Check if the visitor has clicked the link before. If the visitor has not
-     * clicked the link before, return true.
-     *
-     * @param Url $url \App\Models\Url
-     */
-    public function isFirstClick(Url $url): bool
-    {
-        $hasVisited = $url->visits()
-            ->where('user_uid', $this->userService->signature())
-            ->exists();
-
-        return $hasVisited ? false : true;
     }
 
     /**
